+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+3.1.0 (2021-05-10)
+------------------
+* Removed redundant argument `capacity` from init `PlotCurve()`. It is
+  not using a ringbuffer and, hence, does not need a capacity.
+
 3.0.1 (2020-08-07)
 ------------------
 Bug-fixes: